Move: import logic to services
